Revert "update: organize services imports by indexing"
This reverts commit 5a1bc7342dedd63084e8d6a38bc8ff5ad2413283.
