--- src/services/index.js
@@ -1,3 +1,0 @@
-export * from './db-service.js'
-export * from './validation-service.js'
-export * from './intersection-service.js'
